+++ core/dbt/include/global_project/macros/schema_tests/relationships.sql
@@ -25,6 +25,6 @@ where parent.to_field is null
 
 
 {% test relationships(model, column_name, to, field) %}
-    {% set macro = adapter.dispatch('test_relationships') %}
+    {% set macro = adapter.dispatch('test_relationships', 'dbt') %}
     {{ macro(model, column_name, to, field) }}
 {% endtest %}